-fix tests execution
